+++ src/cli/commands/keys.ts
@@ -15,26 +15,40 @@ export async function listKeys(
   if (!ensureOpen(container)) return false;
   if (!ensureMaxParameters(params, 1)) return false;
 
-  let key: KVKey | KVQuery | null = null;
+  let query: KVQuery | null;
   if (hasParameter(params, 0)) {
-    key = KVKeyInstance.parse(params[0], true); // Allow range
+    // Validate query
+    try {
+      const parsedKey = KVKeyInstance.parse(params[0], true);
+      query = new KVKeyInstance(parsedKey, true).get();
+    } catch (e) {
+      console.error(`Could not parse query: ${e.message}`);
+      return false;
+    }
   } else {
-    key = null;
+    query = null;
   }
 
-  const childKeys = container.db?.listKeys(key);
+  const childKeys = container.db?.listKeys(query);
+
+  console.log("");
 
   if (childKeys && childKeys.length > 0) {
-    console.log(Colors.bold("Child Keys:"));
+    if (query === null) {
+      console.log(Colors.bold("Root Keys:"));
+    } else {
+      console.log(Colors.bold("Child Keys:"));
+    }
     for (const childKey of childKeys) {
       console.log(`  ${childKey}`);
     }
   } else {
-    if (key === null) {
+    if (query === null) {
       console.log(Colors.yellow("No root keys found."));
     } else {
       console.log(Colors.yellow("No child keys found for the specified key."));
     }
   }
+  console.log("");
   return true;
 }