+++ src/cli/commands/list.ts
@@ -5,6 +5,7 @@ import {
   type KVDBContainer,
 } from "../common.ts";
 import { KVKeyInstance, type KVQuery } from "../../lib/key.ts";
+import { printTransaction } from "../common.ts";
 
 export async function list(
   container: KVDBContainer,
@@ -16,7 +17,13 @@ export async function list(
   let query: KVQuery;
   if (hasParameter(params, 0)) {
     // Validate query
-    query = new KVKeyInstance(KVKeyInstance.parse(params[0], true)).get();
+    try {
+      const parsedKey = KVKeyInstance.parse(params[0], true);
+      query = new KVKeyInstance(parsedKey, true).get();
+    } catch (e) {
+      console.error(`Could not parse query: ${e.message}`);
+      return false;
+    }
   } else {
     console.error("No query supplied.");
     return false;
@@ -24,11 +31,9 @@ export async function list(
 
   console.log("");
 
-  // Iterate over matching entries
-  for await (const entry of container.db!.iterate(query, 100)) {
-    // Display key information
-    const key = new KVKeyInstance(entry.key).stringify();
-    console.log(key, entry.data);
+  // Iterate over matching entries (max 1000)
+  for await (const entry of container.db!.iterate(query, 1000)) {
+    printTransaction(entry);
   }
 
   console.log(""); // Extra newline for separation